Refactor code to initialize clients
