+++ packages/util/src/config.ts
@@ -2,13 +2,17 @@
 // Copyright 2021 Vulcanize, Inc.
 //
 
+import assert from 'assert';
 import fs from 'fs-extra';
 import path from 'path';
 import toml from 'toml';
 import debug from 'debug';
 import { ConnectionOptions } from 'typeorm';
+import { getDefaultProvider } from 'ethers';
 
-import { Config as CacheConfig } from '@vulcanize/cache';
+import { Config as CacheConfig, getCache } from '@vulcanize/cache';
+import { EthClient } from '@vulcanize/ipld-eth-client';
+import { BaseProvider } from '@ethersproject/providers';
 
 const log = debug('vulcanize:config');
 
@@ -64,4 +68,51 @@ export const getConfig = async (configFile: string): Promise<Config> => {
   log('config', JSON.stringify(config, null, 2));
 
   return config;
+};
+
+export const initClients = async (config: Config): Promise<{
+  dbConfig: ConnectionOptions,
+  serverConfig: ServerConfig,
+  upstreamConfig: UpstreamConfig,
+  jobQueueConfig: JobQueueConfig
+  ethClient: EthClient,
+  postgraphileClient: EthClient,
+  ethProvider: BaseProvider
+}> => {
+  const { database: dbConfig, upstream: upstreamConfig, server: serverConfig, jobQueue: jobQueueConfig } = config;
+
+  assert(serverConfig, 'Missing server config');
+  assert(dbConfig, 'Missing database config');
+  assert(upstreamConfig, 'Missing upstream config');
+
+  const { ethServer: { gqlApiEndpoint, gqlPostgraphileEndpoint, rpcProviderEndpoint }, cache: cacheConfig } = upstreamConfig;
+
+  assert(gqlApiEndpoint, 'Missing upstream ethServer.gqlApiEndpoint');
+  assert(gqlPostgraphileEndpoint, 'Missing upstream ethServer.gqlPostgraphileEndpoint');
+  assert(rpcProviderEndpoint, 'Missing upstream ethServer.rpcProviderEndpoint');
+
+  const cache = await getCache(cacheConfig);
+
+  const ethClient = new EthClient({
+    gqlEndpoint: gqlApiEndpoint,
+    gqlSubscriptionEndpoint: gqlPostgraphileEndpoint,
+    cache
+  });
+
+  const postgraphileClient = new EthClient({
+    gqlEndpoint: gqlPostgraphileEndpoint,
+    cache
+  });
+
+  const ethProvider = getDefaultProvider(rpcProviderEndpoint);
+
+  return {
+    dbConfig,
+    serverConfig,
+    upstreamConfig,
+    jobQueueConfig,
+    ethClient,
+    postgraphileClient,
+    ethProvider
+  };
 };
